added text,, techically finished
although practically finished, locations page has been the most troublesome as I could not figure out what was causing it to not align properly. Wanted to try and do something new that i havent done before which were the hover effects, but after doing so, it ruined the alignment i had placed.